backend/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.1
eventlet==0.36.1
PyJWT==2.8.0
